+++ dgraph/cmd/graphql/schema/gqlschema.go
@@ -78,9 +78,9 @@ func validateSchema(schema *ast.SchemaDocument) gqlerror.List {
 	return errs
 }
 
-// GenerateCompleteSchema generates all the required query/mutation/update functions
+// generateCompleteSchema generates all the required query/mutation/update functions
 // for all the types mentioned the the schema.
-func GenerateCompleteSchema(sch *ast.Schema) {
+func generateCompleteSchema(sch *ast.Schema) {
 
 	extenderMap := make(map[string]*ast.Definition)
