+++ dgraph/cmd/graphql/schema/schemagen_test.yml
@@ -1,18 +1,5 @@
 schemas:
   -
-    name: "Scalar data type"
-    input: |
-      type P {
-        id: ID!
-        q: String!
-      }
-    output: |
-      type P {
-        P.q: string
-      }
-      P.q: string .
-
-  -
     name: "Object data type"
     input: |
       type A {
@@ -70,7 +57,6 @@ schemas:
         s: DateTime
         t: Float
         u: ID
-        v: [Int!]!
       }
     output: |
       type X {
@@ -79,11 +65,9 @@ schemas:
         X.r: string
         X.s: dateTime
         X.t: float
-        X.v: [int]
       }
       X.p: int .
       X.q: bool .
       X.r: string .
       X.s: dateTime .
       X.t: float .
-      X.v: [int] .
